Created IArticulateSearchPlugin interface used as service into search controller. Using that approach, we can use DI pattern to plug in any implementation. If no DI patterns will be used, a default implementation will be used [ArticulateSearch.cs]
